requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.0
asynctest==0.13.0
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
dateparser==1.1.0
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
pip==21.3.1
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.2
requests==2.26.0
setuptools==58.5.2
six==1.16.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
ujson==4.2.0
urllib3==1.26.7
yarl==1.7.2
PyYAML==6.0